Refactor league into it's own class
