+++ lib/stat_tracker.rb
@@ -3,336 +3,35 @@ require_relative 'game'
 require_relative 'team'
 require_relative 'game_teams'
 require_relative 'game_team_stats'
+require_relative 'league_stats'
+# require_relative 'csv_reader'
 require 'pry'
 
 class StatTracker
   attr_reader :games, :teams, :game_teams, :games_array
 
   def initialize(locations)
-    @game_data = CSV.read(locations[:games], headers:true,
-       header_converters: :symbol)
-    @team_data = CSV.read(locations[:teams], headers:true,
-       header_converters: :symbol)
-    @game_team_data = CSV.read(locations[:game_teams],
-       headers:true, header_converters: :symbol)
+    # @game_data = CSV.read(locations[:games], headers:true,
+    #    header_converters: :symbol)
+    # @team_data = CSV.read(locations[:teams], headers:true,
+    #    header_converters: :symbol)
+    # @game_team_data = CSV.read(locations[:game_teams],
+    #    headers:true, header_converters: :symbol)
     # @games_array = []
-    @games = Game.fill_game_array(@game_data)
-    @teams = Team.fill_team_array(@team_data)
-    @game_teams = GameTeams.fill_game_teams_array(@game_team_data)
+    # @games = Game.fill_game_array(@game_data)
+    # @teams = Team.fill_team_array(@team_data)
+    # @game_teams = GameTeams.fill_game_teams_array(@game_team_data)
     @game_team_stats = GameTeamStats.new(locations)
+    @league_stats = LeagueStats.new(locations)
   end
 
   def self.from_csv(locations)
     StatTracker.new(locations)
   end
 
-# Game Statistics
-  def highest_total_score
-    highest_sum = 0
-    @games.each do |game|
-      sum = game.away_goals.to_i + game.home_goals.to_i
-      highest_sum = sum if sum > highest_sum
-    end
-    highest_sum
-  end
-
-  def lowest_total_score
-    lowest_sum = 0
-    lowest = @games.map do |game|
-      sum = game.away_goals.to_i + game.home_goals.to_i
-    end
-    lowest.min
-  end
-
-  def percentage_home_wins
-    home_wins = []
-    @game_teams.each do |game|
-      if game.hoa == "home" && game.result == "WIN"
-        home_wins << game
-      end
-     end
-     (home_wins.length.to_f / @games.length.to_f).round(2)
 
 
-  end
 
-  def percentage_visitor_wins
-    away_wins = []
-    @game_teams.each do |game|
-      if game.hoa == "away" && game.result == "WIN"
-        away_wins << game
-      end
-     end
-     (away_wins.length.to_f / @games.length.to_f).round(2)
-     # require 'pry'; binding.pry
-  end
-
-  def percentage_ties
-    # require 'pry'; binding.pry
-    ((@game_teams.find_all { |game| game.result == 'TIE'}.count) / @game_teams.count.to_f).round(2)
-  end 
-
-  def count_of_games_by_season
-    # require 'pry'; binding.pry
-    @games.group_by {|game| game.season}.transform_values { |values| values.count}
-  end
-
-  def average_goals_per_game
-    avg_goals = []
-    @games.each do |game|
-      goals = game.away_goals + game.home_goals
-      avg_goals << goals
-    end
-    (avg_goals.sum.to_f / avg_goals.length).round(2)
-  end
-
-  def average_goals_by_season
-    # season_goals = {}
-    # season_goals = @games.group_by do |game|
-    #   game.season
-    # end
-    # new_season_goals = season_goals.map do |season, games|
-    #   games.map! do |game|
-    #     game.away_goals + game.home_goals
-    #   end.sum
-      # require 'pry' ; binding.pry
-    # end
-    avg_goals_hash = {}
-    avg = @games.group_by {|game| game.season}
-    avg.map { |season, games| games.map!{ |game| game.away_goals + game.home_goals } }
-    avg.map { |season, games| avg_goals_hash[season] = (games.sum / games.length.to_f).round(2) }
-    avg_goals_hash
-  end
-
-
-# League Statistics
-  def count_of_teams
-    @teams.count
-  end
-
-  def best_offense
-    team_hash = {}
-    @game_teams.each do |game_team|
-      if team_hash[game_team.team_id].nil?
-        team_hash[game_team.team_id] = [game_team.goals]
-      else
-        team_hash[game_team.team_id] << game_team.goals
-      end
-    end
-    sum_goals = 0
-      team_hash.map do |team, goals|
-        goals.each do |goal|
-          sum_goals += goal
-        end
-        avg_goals = sum_goals.to_f / goals.length.to_f
-        team_hash[team] = avg_goals
-        sum_goals = 0
-      end
-      team_hash = team_hash.sort_by {|team_id, avg_goals| avg_goals}.reverse.to_h
-
-      highest_avg = team_hash.values[0]
-      team_hash.map do |team, avg_goals|
-        team_hash.delete(team) if avg_goals < highest_avg
-      end
-
-      if team_hash.length > 1
-        team_names = []
-        team_hash.keys.each do |team_id|
-          team_names << team_name_helper(team_id)
-        end
-        return team_names
-      else
-        return team_name_helper(team_hash.keys[0])
-      end
-  end
-
-  def worst_offense
-    team_hash = {}
-    @game_teams.each do |game_team|
-      if team_hash[game_team.team_id].nil?
-        team_hash[game_team.team_id] = [game_team.goals]
-      else
-        team_hash[game_team.team_id] << game_team.goals
-      end
-    end
-    sum_goals = 0
-      team_hash.map do |team, goals|
-        goals.each do |goal|
-          sum_goals += goal
-        end
-        avg_goals = sum_goals.to_f / goals.length.to_f
-        team_hash[team] = avg_goals
-        sum_goals = 0
-      end
-      team_hash = team_hash.sort_by {|team_id, avg_goals| avg_goals}.to_h
-
-      lowest_avg = team_hash.values[0]
-      team_hash.map do |team, avg_goals|
-        team_hash.delete(team) if avg_goals > lowest_avg
-      end
-
-      if team_hash.length > 1
-        team_names = []
-        team_hash.keys.each do |team_id|
-          team_names << team_name_helper(team_id)
-        end
-        return team_names
-      else
-        return team_name_helper(team_hash.keys[0])
-      end
-  end
-
-  def lowest_scoring_visitor
-    team_hash = {}
-    @games.each do |game|
-      if team_hash[game.away_team_id].nil?
-        team_hash[game.away_team_id] = [game.away_goals]
-      else
-        team_hash[game.away_team_id] << game.away_goals
-      end
-    end
-    sum_goals = 0
-      team_hash.map do |team, goals|
-        goals.each do |goal|
-          sum_goals += goal
-        end
-        avg_goals = sum_goals.to_f / goals.length.to_f
-        team_hash[team] = avg_goals
-        sum_goals = 0
-      end
-      team_hash = team_hash.sort_by {|away_team_id, avg_goals| avg_goals}.to_h
-
-      lowest_avg = team_hash.values[0]
-      team_hash.map do |team, avg_goals|
-        team_hash.delete(team) if avg_goals > lowest_avg
-      end
-
-      if team_hash.length > 1
-        team_names = []
-        team_hash.keys.each do |away_team_id|
-          team_names << team_name_helper(away_team_id)
-        end
-        return team_names
-      else
-        return team_name_helper(team_hash.keys[0])
-      end
-  end
-
-  def highest_scoring_visitor
-    team_hash = {}
-    @games.each do |game|
-      if team_hash[game.away_team_id].nil?
-        team_hash[game.away_team_id] = [game.away_goals]
-      else
-        team_hash[game.away_team_id] << game.away_goals
-      end
-    end
-    sum_goals = 0
-      team_hash.map do |team, goals|
-        goals.each do |goal|
-          sum_goals += goal
-        end
-        avg_goals = sum_goals.to_f / goals.length.to_f
-        team_hash[team] = avg_goals
-        sum_goals = 0
-      end
-      team_hash = team_hash.sort_by {|away_team_id, avg_goals| avg_goals}.reverse.to_h
-
-      highest_avg = team_hash.values[0]
-      team_hash.map do |team, avg_goals|
-        team_hash.delete(team) if avg_goals < highest_avg
-      end
-
-      if team_hash.length > 1
-        team_names = []
-        team_hash.keys.each do |away_team_id|
-          team_names << team_name_helper(away_team_id)
-        end
-        return team_names
-      else
-        return team_name_helper(team_hash.keys[0])
-      end
-  end
-
-  def highest_scoring_home_team
-    team_hash = {}
-    @games.each do |game|
-      if team_hash[game.home_team_id].nil?
-        team_hash[game.home_team_id] = [game.home_goals]
-      else
-        team_hash[game.home_team_id] << game.home_goals
-      end
-    end
-    sum_goals = 0
-      team_hash.map do |team, goals|
-        goals.each do |goal|
-          sum_goals += goal
-        end
-        avg_goals = sum_goals.to_f / goals.length.to_f
-        team_hash[team] = avg_goals
-        sum_goals = 0
-      end
-      team_hash = team_hash.sort_by {|away_team_id, avg_goals| avg_goals}.reverse.to_h
-
-      highest_avg = team_hash.values[0]
-      team_hash.map do |team, avg_goals|
-        team_hash.delete(team) if avg_goals < highest_avg
-      end
-
-      if team_hash.length > 1
-        team_names = []
-        team_hash.keys.each do |away_team_id|
-          team_names << team_name_helper(away_team_id)
-        end
-        return team_names
-      else
-        return team_name_helper(team_hash.keys[0])
-      end
-  end
-
-  def lowest_scoring_home_team
-    team_hash = {}
-    @games.each do |game|
-      if team_hash[game.home_team_id].nil?
-        team_hash[game.home_team_id] = [game.home_goals]
-      else
-        team_hash[game.home_team_id] << game.home_goals
-      end
-    end
-    sum_goals = 0
-      team_hash.map do |team, goals|
-        goals.each do |goal|
-          sum_goals += goal
-        end
-        avg_goals = sum_goals.to_f / goals.length.to_f
-        team_hash[team] = avg_goals
-        sum_goals = 0
-      end
-      team_hash = team_hash.sort_by {|away_team_id, avg_goals| avg_goals}.to_h
-
-      lowest_avg = team_hash.values[0]
-      team_hash.map do |team, avg_goals|
-        team_hash.delete(team) if avg_goals > lowest_avg
-      end
-
-      if team_hash.length > 1
-        team_names = []
-        team_hash.keys.each do |away_team_id|
-          team_names << team_name_helper(away_team_id)
-        end
-        return team_names
-      else
-        return team_name_helper(team_hash.keys[0])
-      end
-  end
-
-  def team_name_helper(team_id)
-    @teams.each do |team|
-      if team.team_id == team_id
-        return team.team_name
-      end
-    end
-  end
 
   # Season Statistics
 
@@ -493,5 +192,33 @@ class StatTracker
 
   def rival(team_id)
     @game_team_stats.rival(team_id)
+  end
+
+  def count_of_teams
+    @league_stats.count_of_teams
+  end
+
+  def best_offense
+    @league_stats.best_offense
+  end
+
+  def worst_offense
+    @league_stats.worst_offense
+  end
+
+  def highest_scoring_visitor
+    @league_stats.highest_scoring_visitor
+  end
+
+  def lowest_scoring_visitor
+    @league_stats.lowest_scoring_visitor
+  end
+
+  def lowest_scoring_home_team
+    @league_stats.lowest_scoring_home_team
+  end
+
+  def highest_scoring_home_team
+    @league_stats.highest_scoring_home_team
   end
 end
