+++ spec/stat_tracker_spec.rb
@@ -26,49 +26,6 @@ RSpec.describe StatTracker do
         expect(@stat_tracker.game_teams.count).to eq(40)
     end
 
-    # xit 'can load collections' do
-    #
-    #   expect(stat_tracker.load_collections(locations)).to eq({
-    #     games => CSV.read(locations[:games], headers:true,
-    #        header_converters: :symbol),
-    #     teams =>,
-    #     game_teams =>
-    #     })
-    #   collections is the key with the data as the value so that
-    # end
-    describe 'League Statistics' do
-      it '#count_of_teams can count teams' do
-      expect(@stat_tracker.count_of_teams).to eq(32)
-      end
-
-      it '#best_offense finds team with the best offense' do
-        expect(@stat_tracker.best_offense).to eq(["Atlanta United",
-           "Orlando City SC", "Portland Timbers", "San Jose Earthquakes"])
-      end
-
-      it '#worst_offense finds team with the worst offense' do
-        expect(@stat_tracker.worst_offense).to eq("Seattle Sounders FC")
-      end
-
-      it '#lowest_scoring_visitor finds team with the lowest average away goals' do
-        expect(@stat_tracker.lowest_scoring_visitor).to eq("Sporting Kansas City")
-      end
-
-      it '#lowest_scoring_home_team finds team with the lowest average home goals' do
-        expect(@stat_tracker.lowest_scoring_home_team).to eq("Seattle Sounders FC")
-      end
-
-      it '#highest_scoring_home_team finds team with the highest average home goals' do
-        expect(@stat_tracker.highest_scoring_home_team).to eq(["Orlando City SC", "San Jose Earthquakes"])
-      end
-
-      it '#team_name_helper finds team name via team_id' do
-        expect(@stat_tracker.team_name_helper("3")).to eq("Houston Dynamo")
-      end
-
-
-    end
-
     describe 'Season Statistics' do
 
       it "organizes seasons by year" do
